backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
google-auth>=2.37.0
google-genai>=1.65.0
redis>=5.2.0
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
